\name{testdata}
\Rdversion{1.1}
\alias{testdata}
\docType{data}
\title{
Real metagenomic data
}
\description{
The metagenomic dataset was downloaded from dbGaP under study ID phs000258. The data and analytical results were first reported by Zupancic et al. (2012). There were a total of 310 Amish adult samples with 112 males and 198 females. And there were a total of 240 taxa at the genus level.
}
\usage{data(testdata)}
\format{
testdata is a data frame with 310 cases(rows) and 248 variables(columns). Among 248 variables, 240 of them are taxa at the genus level and 8 of them are clinical phenotypes.
}
\keyword{datasets}
